Move V8 platform setup to lo.cc
Moves over V8 platform intialization to lo.cc. Platform is initialized via `lo::Setup` and disposed via `lo_cleanup`. This makes it slightly easier to add multi-engine support.
