+++ main.cc
@@ -8,30 +8,6 @@
 #include <fcntl.h>
 #include <sys/mount.h>
 
-using v8::V8;
-using v8::Platform;
-
-// keep the /dev/urandom file open for lifetime of process
-int random_fd = -1;
-
-/**
- * fill the provided buffer with random bytes
- * 
- * we can just use /dev/urandom here, like v8 already does, or come up
- * with something better. it would be nice if we could do this from the
- * JS side, but that doesn't seem possible right now
- * 
- * @param buffer Write random bytes in here.
- * @param length Write this number of random bytes, no more, no less.
- */
-bool EntropySource(unsigned char* buffer, size_t length) {
-  if (random_fd == -1) random_fd = open("/dev/urandom", O_RDONLY);
-  //todo check return
-  size_t bytes = read(random_fd, buffer, length);
-  if (bytes != length) return false;
-  return true;
-}
-
 int main(int argc, char** argv) {
   // if we are called with no arguments, just dump the version and exit
   if (argc == 2 && strncmp(argv[1], "--version", 9) == 0) {
@@ -45,39 +21,16 @@ int main(int argc, char** argv) {
   // https://en.cppreference.com/w/c/io/setvbuf
   setvbuf(stdout, nullptr, _IONBF, 0);
   setvbuf(stderr, nullptr, _IONBF, 0);
-  // create the v8 platform
-  std::unique_ptr<Platform> platform = 
-    v8::platform::NewDefaultPlatform(_v8_threads, 
-      v8::platform::IdleTaskSupport::kDisabled, 
-      v8::platform::InProcessStackDumping::kDisabled, nullptr);
-  V8::InitializePlatform(platform.get());
-  // set the v8 flags from the internally defined ones
-  V8::SetFlagsFromString(v8flags);
-  // then any flags specified on command line will override these, if we 
-  // allow this
-  if (_v8flags_from_commandline == 1) {
-    V8::SetFlagsFromCommandLine(&argc, argv, true);
-  }
-  // V8 requires an entropy source - by default it opens /dev/urandom multiple
-  // times on startup, which we want to avoid. so we need to see if we can
-  // find a more efficient way of providing entropy at startup
-  V8::SetEntropySource(EntropySource);
-  V8::Initialize();
-  V8::InitializeICU();
+
+  lo::Setup(argc, argv, v8flags, _v8_threads, _v8flags_from_commandline);
+
   // register any builtins and modules that have been generated in main.h 
   register_builtins();
   // create a new isolate on the main thread. this will block until the 
   // isolate exits
   lo::CreateIsolate(argc, argv, main_js, main_js_len, index_js, index_js_len, 0,
     0, 0, starttime, RUNTIME, "main.js", _v8_cleanup, _on_exit, nullptr);
-  // if we have the cleanup flag set, clean up memory left behind when isolate
-  // exits. this flag should be set if you want to spawn multiple isolates
-  // in the same process without memory leaks.
-  if (_v8_cleanup) {
-    V8::Dispose();
-    platform.reset();
-  }
-  close(random_fd);
-  lo_shutdown();
+
+  lo_shutdown(_v8_cleanup);
   return 0;
 }
